Disclaimer: I probably missed a bunch of stuff, but I think I hit most of the important points sorry if I wrote an essay, wanted to make sure I covered everything
Fixed some errors in DUMBClient, but not all
Deleted empty (int) and lock (int) fields from struct messageBox,
and implemented mutexes (have not tested it yet, but it makes and runs)
Also updated the signal handler to close (join) all threads when the server closes (not sure if it works)
Not the most necessary, but for errorChecker, some of the same errors print different strings
For example, OPNBX returns ER:OPEND which prints "A box with that name is currently opened by another user, so you can not open it"
while DELBX returns ER:OPEN which prints "A box with that name is currently open, so you can not delete it"
Moved pop() to DUMBlib.c (not tested yet)
Completed the PUTMG function in the server side, but I have not parsed the string from the client.
For OPNBX, I added the ER:BXOPN that you wanted for the extra credit.

Before changing anything, make sure to run the following...
Client: create -> mybox -> open -> mybox -> close -> mybox -> delete -> mybox
It should work, I tested it multiple times. However, there are still a bunch of client-side string errors

For example, try running: create -> mybox -> create -> mybox2
It seg faults on the client, and blows up the server

What needs to be done:
1. add BXOPN error to errorChecker
2. fix string stuff on client.c
3. PUTMG string parsing on the server side
4. NEXTMG needs to be tested once PUTMG is done
5. Testplan (we don't need a readme)
6. Converting hostname into IP (I will 100% ask Ben tomorrow)
7. Also, if you can, output a string to prompt the user to input the box name (the example on the asst description shows the client outputting a bunch of stuff. Idk if it's necessary though)